requirements.txt
fire
requests
pillow
opencv-python
pycocotools
